+++ client/src/App.tsx
@@ -1,9 +1,8 @@
 import { lazy, Suspense, useEffect } from "react";
 import { Routes, Route } from "react-router-dom";
 
-import NotFound from "@/components/layout/NotFound";
-
 import LoadingPage from "@/pages/Loading.tsx";
+import NotFound from "@/pages/NotFound";
 
 import { denamuAscii } from "@/constants/denamuAscii.ts";
